Refactored interfaces and adapters
