+++ IndyPOS.sln
@@ -10,17 +10,27 @@ EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Debug|x64.ActiveCfg = Debug|x64
+		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Debug|x64.Build.0 = Debug|x64
 		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Release|x64.ActiveCfg = Release|x64
+		{2ED6DFC3-BF73-41B7-816C-D53312DA9C52}.Release|x64.Build.0 = Release|x64
 		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Debug|x64.ActiveCfg = Debug|x64
+		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Debug|x64.Build.0 = Debug|x64
 		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Release|x64.ActiveCfg = Release|x64
+		{6032F51E-E3DF-47D5-9E74-A360BDBFB180}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
