+++ IndyPOS/UI/UpdateInventoryProductForm.cs
@@ -1,6 +1,7 @@
 ﻿using IndyPOS.Adapters;
 using IndyPOS.Constants;
 using IndyPOS.Controllers;
+using IndyPOS.Inventory;
 using Prism.Events;
 using System;
 using System.Collections.Generic;
